Refactoring a bit AerialView in order to implement some new features
